Move re-broadcasting functionality from actor ctor to housekeeping
