+++ daemon/src/taker.rs
@@ -156,6 +156,9 @@ async fn main() -> Result<()> {
                 housekeeping::transition_non_continue_cfds_to_setup_failed(&mut conn)
                     .await
                     .unwrap();
+                housekeeping::rebroadcast_transactions(&mut conn, &wallet)
+                    .await
+                    .unwrap();
                 let cfds = load_all_cfds(&mut conn).await.unwrap();
 
                 let (cfd_feed_sender, cfd_feed_receiver) = watch::channel(cfds.clone());
@@ -164,7 +167,6 @@ async fn main() -> Result<()> {
                     .spawn_global();
 
                 let (monitor_actor_address, mut monitor_actor_context) = xtra::Context::new(None);
-
                 let (oracle_actor_address, mut oracle_actor_context) = xtra::Context::new(None);
 
                 let mut conn = db.acquire().await.unwrap();
@@ -177,11 +179,8 @@ async fn main() -> Result<()> {
                     order_feed_sender,
                     send_to_maker,
                     monitor_actor_address.clone(),
-                    cfds.clone(),
                     oracle_actor_address,
                 )
-                .await
-                .unwrap()
                 .create(None)
                 .spawn_global();
